clean up, definition of env file for keys
